[refactor]ドメイン駆動設計のためにフォルダ整理
