+++ server/nodemon.json
@@ -1,7 +1,5 @@
 {
-  "watch": [
-    "dist"
-  ],
+  "watch": ["dist"],
   "ext": "js",
-  "exec": "node dist/src/main"
+  "exec": "node dist/main"
 }
