+++ server/package.json
@@ -7,10 +7,10 @@
   "scripts": {
     "build": "tsc -p tsconfig.build.json",
     "format": "prettier --write \"src/**/*.ts\"",
-    "migration:up": "ts-node $(npm bin)/typeorm migration:run",
+    "migration:up": "ts-node $(npm bin)/typeorm migration:run -f src/infrastructure/config/ormconfig.js",
     "migration:revert": "ts-node $(npm bin)/typeorm migration:revert",
     "start": "ts-node -r tsconfig-paths/register src/main.ts",
-    "start:dev": "concurrently --handle-input \"wait-on dist/src/main.js && nodemon\" \"tsc -w -p tsconfig.build.json\" ",
+    "start:dev": "concurrently --handle-input \"wait-on dist/main.js && nodemon\" \"tsc -w -p tsconfig.build.json\" ",
     "start:debug": "nodemon --config nodemon-debug.json",
     "prestart:prod": "rimraf dist && npm run build",
     "start:prod": "node dist/src/main.js",
