+++ server/src/infrastructure/auth-user/auth-user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { AuthUserController } from './controllers/auth-user.controller';
+import { AuthUserController } from '../../application/auth-user/auth-user.controller';
 import { GoogleStrategy } from './strategies/google.strategy';
-import { AuthUserService } from './services/auth-user.service';
+import { AuthUserService } from '../../domain/auth-user/services/auth-user.service';
 import { JwtStrategy } from './strategies/jwt.strategy';
 
 @Module({
