+++ server/src/infrastructure/auth-user/strategies/jwt.strategy.ts
@@ -1,9 +1,8 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { EnvManager } from '../../utils/env.manager';
-import { JwtPayload } from '../domains/jwt.payload';
-import { GoogleProfilesData } from '../../tempData/google.profiles.data';
+import { JwtPayload } from '../../../domain/auth-user/jwt.payload';
+import { EnvManager } from '../../../utils/env.manager';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
