+++ server/tsconfig.json
@@ -11,7 +11,6 @@
     "allowJs": true,
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./",
     "incremental": true
   },
   "include": ["src/**/*"],
